move Dialog utils into single class
